[networks] Add missing build tag
