+++ pkg/process/util/util.go
@@ -123,17 +123,6 @@ func GetSysRoot() string {
 	return "/sys"
 }
 
-// GetRootNSPID returns the current PID from the root namespace
-func GetRootNSPID() (int, error) {
-	pidPath := filepath.Join(GetProcRoot(), "self")
-	pidStr, err := os.Readlink(pidPath)
-	if err != nil {
-		return 0, err
-	}
-
-	return strconv.Atoi(pidStr)
-}
-
 // WithAllProcs will execute `fn` for every pid under procRoot. `fn` is
 // passed the `pid`. If `fn` returns an error the iteration aborts,
 // returning the last error returned from `fn`.
