+++ pkg/process/util/util_linux.go
@@ -1,0 +1,20 @@
+// +build linux
+
+package util
+
+import (
+	"os"
+	"path/filepath"
+	"strconv"
+)
+
+// GetRootNSPID returns the current PID from the root namespace
+func GetRootNSPID() (int, error) {
+	pidPath := filepath.Join(GetProcRoot(), "self")
+	pidStr, err := os.Readlink(pidPath)
+	if err != nil {
+		return 0, err
+	}
+
+	return strconv.Atoi(pidStr)
+}
