+++ pkg/process/util/util_linux_test.go
@@ -1,3 +1,5 @@
+// +build linux
+
 package util
 
 import (
